+++ src/Resources/app/administration/src/component/slide/preview/cover/index.ts
@@ -67,8 +67,4 @@ export default Component.wrapComponentConfig({
             return useViewportProp(property, this.deviceView, this.slide.slideSettings.viewports)
         }
     },
-
-    created() {
-        console.log(this.slide, this.deviceView)
-    }
 });