+++ src/Resources/app/administration/src/extension/sw-cms-sidebar/index.ts
@@ -10,5 +10,18 @@ export default Component.wrapComponentConfig({
         blurElysiumSectionName() {
             return 'blur-elysium-section';
         },
+
+        elysiumSectionSettingsTitle() {
+            if (this.selectedBlock?.customFields?.hasOwnProperty('elysiumBlockSettings')) {
+                return this.$tc('blurElysiumSection.sidebar.blockSettingsTitle');
+            } else {
+                return this.$tc('blurElysiumSection.sidebar.sectionSettingsTitle');
+            }
+        },
+
+        elysiumSectionSettingsActive() {
+            return this.selectedSection?.customFields?.hasOwnProperty('elysiumSectionSettings') ||
+                this.selectedBlock?.customFields?.hasOwnProperty('elysiumBlockSettings');
+        }
     },
 });