+++ src/Resources/app/administration/src/extension/sw-sidebar/index.ts
@@ -1,0 +1,8 @@
+import template from './template.html.twig';
+
+// eslint-disable-next-line no-undef
+const { Component } = Shopware;
+
+export default Component.wrapComponentConfig({
+    template,
+});